move approval-psikiater controller, routes to admins
