+++ routes/admins.js
@@ -3,4 +3,5 @@ const Router = express.Router();
 const AdminController = require("../controllers/admins");
 
 Router.post("/create-admin", AdminController.createAdmin);
+Router.post("/approval", AdminController.approvalPsikiater);
 module.exports = Router;
