--- routes/approval.js
@@ -1,6 +1,0 @@
-const express = require("express");
-const Router = express.Router();
-const ApprovalController = require("../controllers/approval");
-
-Router.post("/approval", ApprovalController.approvalPsikiater);
-module.exports = Router;
